+++ examples/resources/cloudflare_load_balancer_monitor/import.sh
@@ -1,0 +1,1 @@
+$ terraform import cloudflare_load_balancer_monitor.example <account_id>/<load_balancer_monitor_id>